requirements.txt
Pillow
cryptography